Refactor to make the Monthly Day component composable
