+++ .gitignore
@@ -4,3 +4,4 @@ node_modules
 .cache
 dist
 .parcel-cache
+coverage
