+++ src/index.tsx
@@ -1,4 +1,5 @@
 export * from './Monthly/MonthlyCalendar';
+export * from './Monthly/MonthlyBody';
 export * from './Monthly/MonthlyEventItems';
 
 export * from './Weekly/WeeklyCalendar';
